% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.hifld.mental-package.R
\docType{package}
\name{dc.hifld.mental-package}
\alias{dc.hifld.mental}
\alias{dc.hifld.mental-package}
\title{dc.hifld.mental: Social Data Commons: Behavioral Health Facility Access Scores}
\description{
Allows user to easily get Behavioral Health Facility Access Scores data from the SDAD Data Commons.  Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Cesar Montalvo \email{cpm9w@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
